add filters to the order list handler and update order and user id to int64

+++ internal/aaa/bl/storage.go
@@ -1,8 +1,8 @@
 package bl
 
 type UserStorage interface {
-	CheckUserByEmail(email string) (int, int, error)
-	CreateUser(email string) (int, error)
+	CheckUserByEmail(email string) (int64, int, error)
+	CreateUser(email string) (int64, error)
 }
 
 type Storage interface {
